Cast Nova Furry XL pipeline to float32 when moving to CPU

+++ nova_furry_xl/pytorch/src/model_utils.py
@@ -27,7 +27,7 @@ def load_pipe(variant):
     modules = [pipe.text_encoder, pipe.unet, pipe.text_encoder_2, pipe.vae]
 
     # Move the pipeline to CPU
-    pipe.to("cpu")
+    pipe.to("cpu", dtype=torch.float32)
 
     for module in modules:
         module.eval()
